backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
faiss-cpu==1.7.4
openai==1.3.7
azure-openai==1.3.0
numpy==1.24.3
scikit-learn==1.3.2